# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NutriBasket API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /logs/daily:
    get:
      operationId: listDailyLogs
      tags: [meals]
      summary: List daily logs (with date filter)
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of daily logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyLog"
  /logs/daily/{date}:
    get:
      operationId: getDailyLog
      tags: [meals]
      summary: Get a single day's log
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
    put:
      operationId: upsertDailyLog
      tags: [meals]
      summary: Update water intake, weight, or notes for a day
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyLogInput"
      responses:
        "200":
          description: Updated daily log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
  /logs/daily/{date}/meals:
    get:
      operationId: getMealEntries
      tags: [meals]
      summary: Get all meal entries for a day
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Meal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealEntry"
    post:
      operationId: addMealEntry
      tags: [meals]
      summary: Add a meal entry to a day
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealEntryInput"
      responses:
        "201":
          description: Created meal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealEntry"
  /logs/daily/{date}/meals/{id}:
    delete:
      operationId: deleteMealEntry
      tags: [meals]
      summary: Remove a meal entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /logs/activity:
    get:
      operationId: listActivityLogs
      tags: [activity]
      summary: List activity logs
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
    post:
      operationId: createActivityLog
      tags: [activity]
      summary: Log activity for a day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityLogInput"
      responses:
        "201":
          description: Created activity log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLog"
  /logs/activity/{id}:
    put:
      operationId: updateActivityLog
      tags: [activity]
      summary: Update an activity log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityLogInput"
      responses:
        "200":
          description: Updated activity log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLog"
    delete:
      operationId: deleteActivityLog
      tags: [activity]
      summary: Delete an activity log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- Recipes ----
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Search and list products across retailers
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: retailerId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: onSpecial
          in: query
          schema:
            type: boolean
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/compare:
    get:
      operationId: compareProduct
      tags: [products]
      summary: Get the same/similar product across all retailers for price comparison
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cross-retailer comparison
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductComparison"

  # ---- Basket ----
...
components:
  schemas:
...
    DailyLog:
      type: object
      required: [date, totalCalories, totalProteinG, totalCarbsG, totalFatG, waterMl, calorieTarget, proteinTarget, carbsTarget, fatTarget, adherencePercent, streak]
      properties:
        date:
          type: string
          format: date
        totalCalories:
          type: integer
        totalProteinG:
          type: number
        totalCarbsG:
          type: number
        totalFatG:
          type: number
        waterMl:
          type: integer
        calorieTarget:
          type: integer
        proteinTarget:
          type: integer
        carbsTarget:
          type: integer
        fatTarget:
          type: integer
        adherencePercent:
          type: number
        streak:
          type: integer
        weightKg:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    DailyLogInput:
      type: object
      properties:
        waterMl:
          type: integer
        weightKg:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    MealEntry:
      type: object
      required: [id, date, mealType, name, calories, proteinG, carbsG, fatG, servings]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        name:
          type: string
        calories:
          type: integer
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        servings:
          type: number
        recipeId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        isFavourite:
          type: boolean
    MealEntryInput:
      type: object
      required: [mealType, name, calories, proteinG, carbsG, fatG, servings]
      properties:
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        name:
          type: string
        calories:
          type: integer
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        servings:
          type: number
        recipeId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        isFavourite:
          type: boolean
...
    ActivityLog:
      type: object
      required: [id, date, steps, activeCalories, workoutDurationMin, sleepHours, estimatedCaloriesBurned, netCalorieBalance]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        steps:
          type: integer
        activeCalories:
          type: integer
        workoutDurationMin:
          type: integer
        workoutType:
          type: ["string", "null"]
        sleepHours:
          type: number
        estimatedCaloriesBurned:
          type: integer
        netCalorieBalance:
          type: integer
        notes:
          type: ["string", "null"]
    ActivityLogInput:
      type: object
      required: [date, steps, activeCalories, workoutDurationMin, sleepHours]
      properties:
        date:
          type: string
          format: date
        steps:
          type: integer
        activeCalories:
          type: integer
        workoutDurationMin:
          type: integer
        workoutType:
          type: ["string", "null"]
        sleepHours:
          type: number
        notes:
          type: ["string", "null"]

    # ---- Recipes ----
...
    Product:
      type: object
      required: [id, name, retailerId, retailerName, category, priceAud, packSize, packUnit, caloriesPer100g, proteinPer100g, carbsPer100g, fatPer100g, isOnSpecial, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: ["string", "null"]
        retailerId:
          type: integer
        retailerName:
          type: string
        category:
          type: string
          enum: [protein, dairy, pantry, fruit_veg, snacks, drinks, grains, condiments, frozen, other]
        priceAud:
          type: number
        regularPriceAud:
          type: ["number", "null"]
        packSize:
          type: number
        packUnit:
          type: string
        caloriesPer100g:
          type: integer
        proteinPer100g:
          type: number
        carbsPer100g:
          type: number
        fatPer100g:
          type: number
        fiberPer100g:
          type: ["number", "null"]
        sugarPer100g:
          type: ["number", "null"]
        isOnSpecial:
          type: boolean
        savingsPercent:
          type: ["number", "null"]
        savingsAud:
          type: ["number", "null"]
        imageUrl:
          type: string
    ProductComparison:
      type: object
      required: [product, pricePerUnit, isCheapest, isBestValue]
      properties:
        product:
          $ref: "#/components/schemas/Product"
        pricePerUnit:
          type: number
        isCheapest:
          type: boolean
        isBestValue:
          type: boolean

    # ---- Basket ----
...