# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriBasket API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  date: Date;
  steps: number;
  activeCalories: number;
  workoutDurationMin: number;
  /** @nullable */
  workoutType?: string | null;
  sleepHours: number;
  estimatedCaloriesBurned: number;
  netCalorieBalance: number;
  /** @nullable */
  notes?: string | null;
}