# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriBasket API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from './productCategory';

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  brand?: string | null;
  retailerId: number;
  retailerName: string;
  category: ProductCategory;
  priceAud: number;
  /** @nullable */
  regularPriceAud?: number | null;
  packSize: number;
  packUnit: string;
  caloriesPer100g: number;
  proteinPer100g: number;
  carbsPer100g: number;
  fatPer100g: number;
  /** @nullable */
  fiberPer100g?: number | null;
  /** @nullable */
  sugarPer100g?: number | null;
  isOnSpecial: boolean;
  /** @nullable */
  savingsPercent?: number | null;
  /** @nullable */
  savingsAud?: number | null;
  imageUrl: string;
}